# Optimize SlidingTimeWindowMovingAverages sumBuckets()