--- a/tritium-registry/src/main/java/com/palantir/tritium/metrics/registry/OptimizedSlidingTimeWindowMovingAverages.java
+++ b/tritium-registry/src/main/java/com/palantir/tritium/metrics/registry/OptimizedSlidingTimeWindowMovingAverages.java
@@ -27,7 +27,7 @@ import java.time.Duration;
 import java.time.Instant;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
-import java.util.concurrent.atomic.LongAdder;
+import java.util.concurrent.atomic.AtomicLongArray;
 
 /**
  * Optimized fork of {@link com.codahale.metrics.SlidingTimeWindowMovingAverages}, see
@@ -64,7 +64,7 @@ public final class OptimizedSlidingTimeWindowMovingAverages implements MovingAve
      * One counter per time bucket/slot (i.e. per second, see TICK_INTERVAL) for the entire
      * time window (i.e. 15 minutes, see TIME_WINDOW_DURATION_MINUTES).
      */
-    private final LongAdder[] buckets;
+    private final AtomicLongArray buckets;
 
     /**
      * Index into buckets, pointing at the bucket containing the oldest counts.
@@ -104,10 +104,7 @@ public final class OptimizedSlidingTimeWindowMovingAverages implements MovingAve
         final long startTime = clock.getTick();
         this.lastTick = new AtomicLong(startTime);
 
-        this.buckets = new LongAdder[NUMBER_OF_BUCKETS];
-        for (int i = 0; i < NUMBER_OF_BUCKETS; i++) {
-            this.buckets[i] = new LongAdder();
-        }
+        this.buckets = new AtomicLongArray(NUMBER_OF_BUCKETS);
         this.bucketBaseTime = Instant.ofEpochSecond(0L, startTime);
         this.oldestBucketTime = bucketBaseTime;
         this.oldestBucketIndex = 0;
@@ -116,7 +113,7 @@ public final class OptimizedSlidingTimeWindowMovingAverages implements MovingAve
 
     @Override
     public void update(long value) {
-        buckets[currentBucketIndex].add(value);
+        buckets.addAndGet(currentBucketIndex, value);
     }
 
     @Override
@@ -194,14 +191,14 @@ public final class OptimizedSlidingTimeWindowMovingAverages implements MovingAve
     private void cleanBucketRange(int fromIndex, int toIndex) {
         if (fromIndex < toIndex) {
             for (int i = fromIndex; i < toIndex; i++) {
-                buckets[i].reset();
+                buckets.set(i, 0L);
             }
         } else {
             for (int i = fromIndex; i < NUMBER_OF_BUCKETS; i++) {
-                buckets[i].reset();
+                buckets.set(i, 0L);
             }
             for (int i = 0; i < toIndex; i++) {
-                buckets[i].reset();
+                buckets.set(i, 0L);
             }
         }
     }
@@ -215,14 +212,14 @@ public final class OptimizedSlidingTimeWindowMovingAverages implements MovingAve
         long sum = 0;
         if (fromIndex < toIndex) {
             for (int i = fromIndex; i < toIndex; i++) {
-                sum += buckets[i].longValue();
+                sum += buckets.get(i);
             }
         } else {
             for (int i = fromIndex; i < NUMBER_OF_BUCKETS; i++) {
-                sum += buckets[i].longValue();
+                sum += buckets.get(i);
             }
             for (int i = 0; i < toIndex; i++) {
-                sum += buckets[i].longValue();
+                sum += buckets.get(i);
             }
         }
         return sum;