@schlosna 2025-09-03
> @schlosna  it looks like upstream has released a few times in the past week. If you can wait, I might try to get this released upstream first rather than put a fork here in tritium.

Sounds good, will give it a few days for the upstream PR, and I'll try to put together a quick JMH benchmark demonstrating the delta.