@schlosna 2025-09-04
updated with benchmark showing 22x speedup and ~0 allocations