@schlosna 2025-09-16
closing in favor of https://github.com/palantir/tritium/pull/2265 which incorporates upstream fix in https://github.com/dropwizard/metrics/releases/tag/v4.2.37